pyproject.toml
[tool.poetry]
name = "jsnac"
version = "0.1.0"
description = "JSON Schema (for) Network as Code"
authors = ["Andrew Jones <jsnac@commitconfirmed.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^6.0.*"

[tool.poetry.scripts]
jsnac = "jsnac.utils.jsnac_cli:main"

[tool.poetry.dev-dependencies]
coverage = "^7.6.*"
pytest = "^8.3.*"
jsonschema = "^4.23.*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"